Ran into a problem where Map elements were being suddenly nulled out after having been properly assigned earlier. Going to have to look into that more in the immediate future

+++ Weather.cpp
@@ -9,7 +9,7 @@ void Weather::setWeatherMapSize(int size)
   weatherMap.setSize(size);
 }
 
-Map<int>* Weather::getWeatherMap()
+Map<double>* Weather::getWeatherMap()
 {
   return &weatherMap;
 }
